context method on the iterator
